+++ docs/changelog.md
@@ -1,6 +1,7 @@
 XXXX/XX/XX Release 3.9.14
 
 - Hide primitive/complex datatypes and logical models in the validators [#352](https://github.com/ahdis/matchbox/issues/352)
+- Fix handling of the hl7.fhir.uv.extensions packages [#343](https://github.com/ahdis/matchbox/issues/343)
 
 2025/03/05 Release 3.9.13